+++ src/index.js
@@ -1,7 +1,7 @@
 // index.js
 import React from "react";
 import ReactDOM from "react-dom";
-import Router from "./container/Router.jsx";
+import { Router } from "./container/Router.jsx";
 import { BrowserRouter } from "react-router-dom";
 import { Provider } from "react-redux";
 import initStore from "./utils/store.js";
